:camera_flash: enable VerifyTest's DiffPlex
The text comparison behavior of Verify is pluggable. The default behaviour, on failure, is to output both the received and the verified contents as part of the exception. This can be noisy when verifying large strings.

Verify.DiffPlex changes the text compare result to highlighting text differences inline.

In xUnit, you can use the ModuleInitializer attribute to run code before any tests in an assembly are executed.